refactor: added some folders for certain groups of components
