+++ src/components/TodoComponents/AddTodoModal/AddTodoModal.scss
@@ -1,4 +1,4 @@
-@import "../../scss/index.scss";
+@import "@/scss/index.scss";
 
 .modal-overlay {
   position: fixed;
