+++ src/components/TodoComponents/AddTodoModal/AddTodoModal.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 import { format, parse } from "date-fns";
 import DatePicker from "react-datepicker";
 
-import CustomButton from "../CustomButton";
+import CustomButton from "../../UI/CustomButton";
 import { DATE_FORMAT, TIME_FORMAT, TIME_INTERVAL } from "@/consts/dateFormats";
 import { MAX_INPUT_LENGTH } from "@/consts/inputLength";
 import { getMinDate, getMaxDate } from "@/helpers/getDate";
